Optimize chat messages load and fix BG size
* Optimized chat messages loading.

Previously, we let messages accumulate even when
the chat log panel is disabled. When it's enabled,
all the accumulated messages will be printed and popped
until we are left with 17-18 messages as normal.

This commit will always pop extra old messages, regardless
if the panel is enabled or not.

* Make bigSheet background height lowered down a bit to
cover the last message better.